+++ services/dag/src/dag/web/__main__.py
@@ -14,5 +14,6 @@ def main():
 
     app.run(port=8080)
 
+
 if __name__ == '__main__':
     main()
